requirements.txt
selenium==4.33.0
python-dotenv==1.1.0
ollama==0.5.1